i think it  works now, the dark mode
