+++ components/Task.js
@@ -1,9 +1,11 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
 import { View, Text, TouchableOpacity, TextInput, StyleSheet, Alert } from 'react-native';
+import { ThemeContext } from './ThemeContext';
 
 const Task = ({ task, onIncrement, onUpdateText, onReset, onDelete }) => {
   const [editing, setEditing] = useState(false);
   const [text, setText] = useState(task.text);
+  const { theme } = useContext(ThemeContext);
 
   const handleIncrement = () => {
     setEditing(false);
@@ -29,10 +31,10 @@ const Task = ({ task, onIncrement, onUpdateText, onReset, onDelete }) => {
   };
 
   return (
-    <View style={styles.taskContainer}>
+    <View style={[styles.taskContainer, theme.container]}>
       {editing ? (
         <TextInput
-          style={styles.taskTextInput}
+          style={[styles.taskTextInput, { color: theme.text.color, borderColor: theme.text.color }]}
           value={text}
           onChangeText={setText}
           onBlur={handleTextUpdate}
@@ -44,11 +46,11 @@ const Task = ({ task, onIncrement, onUpdateText, onReset, onDelete }) => {
           onPress={() => setEditing(true)}
           onLongPress={handleLongPress}
         >
-          <Text style={styles.taskText}>{text}</Text>
+          <Text style={[styles.taskText, theme.text]}>{text}</Text>
         </TouchableOpacity>
       )}
       <View style={styles.counterContainer}>
-        <Text style={styles.counterText}>{task.count}</Text>
+        <Text style={[styles.counterText, theme.text]}>{task.count}</Text>
         <TouchableOpacity style={styles.tickButton} onPress={handleIncrement}>
           <Text style={styles.tickButtonText}>✓</Text>
         </TouchableOpacity>
@@ -63,7 +65,6 @@ const styles = StyleSheet.create({
     justifyContent: "space-between",
     alignItems: "center",
     padding: 10,
-    backgroundColor: "#f2f2f2",
     marginVertical: 5,
     borderRadius: 5
   },
@@ -92,13 +93,12 @@ const styles = StyleSheet.create({
     fontSize: 16,
     padding: 5,
     borderWidth: 1,
-    borderColor: 'gray',
     borderRadius: 5,
   },
   taskTextContainer: {
     flex: 1,
     justifyContent: 'center',
   },
-})
+});
 
-export default Task
+export default Task;
